refactor: change api module to depend on core
