+++ blink-api/src/main/kotlin/com/jordyma/blink/auth/jwt/user_account/UserAccount.kt
@@ -1,6 +1,6 @@
 package com.jordyma.blink.auth.jwt.user_account
 
-import com.jordyma.blink.user.entity.Role
+import com.jordyma.blink.user.Role
 import org.springframework.security.core.userdetails.User
