+++ blink-api/src/main/kotlin/com/jordyma/blink/auth/service/AuthService.kt
@@ -8,9 +8,9 @@ import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import com.jordyma.blink.auth.dto.request.AppleLoginRequestDto
-import com.jordyma.blink.user.entity.SocialType
-import com.jordyma.blink.user.entity.User
-import com.jordyma.blink.user.repository.UserRepository
+import com.jordyma.blink.user.SocialType
+import com.jordyma.blink.user.User
+import com.jordyma.blink.user.UserRepository
 import com.jordyma.blink.auth.dto.request.KakaoLoginRequestDto
 import com.jordyma.blink.auth.dto.response.AppleDto
 import com.jordyma.blink.auth.dto.response.TokenResponseDto
@@ -21,18 +21,18 @@ import com.jordyma.blink.global.exception.ErrorCode
 import com.jordyma.blink.global.http.api.KakaoAuthApi
 import com.jordyma.blink.global.http.response.OpenKeyListResponse
 import com.jordyma.blink.auth.jwt.user_account.UserAccount
-import com.jordyma.blink.feed.entity.Feed
-import com.jordyma.blink.feed.entity.Source
-import com.jordyma.blink.feed.entity.Status
-import com.jordyma.blink.feed.repository.FeedRepository
-import com.jordyma.blink.folder.entity.Folder
-import com.jordyma.blink.folder.repository.FolderRepository
-import com.jordyma.blink.keyword.entity.Keyword
-import com.jordyma.blink.keyword.repository.KeywordRepository
+import com.jordyma.blink.feed.Feed
+import com.jordyma.blink.feed.Source
+import com.jordyma.blink.feed.Status
+import com.jordyma.blink.feed.FeedRepository
+import com.jordyma.blink.folder.Folder
+import com.jordyma.blink.folder.FolderRepository
+import com.jordyma.blink.keyword.Keyword
+import com.jordyma.blink.keyword.KeywordRepository
 import com.jordyma.blink.logger
-import com.jordyma.blink.user.entity.Role
-import com.jordyma.blink.user_refresh_token.entity.UserRefreshToken
-import com.jordyma.blink.user_refresh_token.repository.UserRefreshTokenRepository
+import com.jordyma.blink.user.Role
+import com.jordyma.blink.user.UserRefreshToken
+import com.jordyma.blink.user.UserRefreshTokenRepository
 import com.nimbusds.jose.JOSEException
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.JWSHeader
