+++ blink-api/src/main/kotlin/com/jordyma/blink/feed/controller/FeedController.kt
@@ -71,7 +71,7 @@ class FeedController(
         feedSummarizeMessageSender.send(summarizeMessage)
 
         val feedIdResponseDto = FeedIdResponseDto(
-            feedId = feed.id
+            feedId = feed.id!!
         )
         return ResponseEntity.ok(feedIdResponseDto)
     }
