+++ blink-api/src/main/kotlin/com/jordyma/blink/image/service/ImageService.kt
@@ -1,7 +1,7 @@
 package com.jordyma.blink.image.service
 
-import com.jordyma.blink.feed.entity.Feed
-import com.jordyma.blink.feed.repository.FeedRepository
+import com.jordyma.blink.feed.Feed
+import com.jordyma.blink.feed.FeedRepository
 import com.jordyma.blink.global.exception.ApplicationException
 import com.jordyma.blink.global.exception.ErrorCode
 import com.jordyma.blink.global.s3.S3Uploader
