+++ blink-core/src/main/kotlin/com.jordyma.blink/feed/impl/FeedRepositoryCustomImpl.kt
@@ -1,20 +1,20 @@
-package com.jordyma.blink.feed.repository.impl
+package com.jordyma.blink.feed.impl
 
-import com.jordyma.blink.feed.entity.Feed
-import com.jordyma.blink.feed.entity.QFeed
-import com.jordyma.blink.feed.entity.QFeed.feed
-import com.jordyma.blink.feed.entity.Status
-import com.jordyma.blink.feed.repository.FeedRepositoryCustom
+import com.jordyma.blink.feed.Feed
+import com.jordyma.blink.feed.Status
+import com.jordyma.blink.feed.FeedRepositoryCustom
+import com.jordyma.blink.feed.QFeed
+import com.jordyma.blink.feed.QFeed.feed
 import com.jordyma.blink.feed.vo.FeedFolderVo
 import com.jordyma.blink.feed.vo.FeedDetailVo
-import com.jordyma.blink.folder.entity.Folder
-import com.jordyma.blink.folder.entity.QFolder
-import com.jordyma.blink.folder.entity.QFolder.folder
-import com.jordyma.blink.folder.entity.QRecommend
-import com.jordyma.blink.keyword.entity.QKeyword
-import com.jordyma.blink.user.entity.QUser
-import com.jordyma.blink.user.entity.QUser.user
-import com.jordyma.blink.user.entity.User
+import com.jordyma.blink.folder.Folder
+import com.jordyma.blink.folder.QFolder
+import com.jordyma.blink.folder.QFolder.folder
+import com.jordyma.blink.keyword.QKeyword
+import com.jordyma.blink.recommend.QRecommend
+import com.jordyma.blink.user.QUser
+import com.jordyma.blink.user.QUser.user
+import com.jordyma.blink.user.User
 import com.querydsl.core.BooleanBuilder
 import com.querydsl.core.types.Projections
 import com.querydsl.jpa.impl.JPAQueryFactory
