+++ blink-core/src/main/kotlin/com.jordyma.blink/feed/vo/FeedFolderVo.kt
@@ -2,7 +2,6 @@ package com.jordyma.blink.feed.vo
 
 import com.jordyma.blink.feed.Feed
 
-
 data class FeedFolderVo(
     val folderId: Long,
     val folderName: String,
