+++ blink-core/src/main/kotlin/com.jordyma.blink/feed/vo/FeedKeywordVo.kt
@@ -1,7 +1,7 @@
 package com.jordyma.blink.feed.vo
 
-import com.jordyma.blink.feed.entity.Feed
-import com.jordyma.blink.keyword.entity.Keyword
+import com.jordyma.blink.feed.Feed
+import com.jordyma.blink.keyword.Keyword
 
 data class FeedKeywordVo(
     val feed: Feed,
