+++ blink-core/src/main/kotlin/com.jordyma.blink/feed/vo/ScoredFeedVo.kt
@@ -1,6 +1,5 @@
 package com.jordyma.blink.feed.vo
-
-import com.jordyma.blink.feed.entity.Feed
+import com.jordyma.blink.feed.Feed
 
 data class ScoredFeedVo(
     val feed: Feed,
