+++ blink-core/src/main/kotlin/com.jordyma.blink/notice/NoticeRepository.kt
@@ -1,0 +1,6 @@
+package com.jordyma.blink.notice
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface NoticeRepository : JpaRepository<Notice, Long> {
+}
